--- a/query/physicalplan/sampler.go
+++ b/query/physicalplan/sampler.go
@@ -16,7 +16,7 @@ type ReservoirSampler struct {
 	size int64
 
 	// reservoir is the set of records that have been sampled. They may vary in schema due to dynamic columns.
-	reservoir []arrow.Record
+	reservoir []sample
 
 	w float64 // w is the probability of keeping a record
 	n int64   // n is the number of rows that have been sampled thus far
@@ -27,7 +27,7 @@ type ReservoirSampler struct {
 func NewReservoirSampler(size int64) *ReservoirSampler {
 	return &ReservoirSampler{
 		size:      size,
-		reservoir: []arrow.Record{},
+		reservoir: make([]sample, 0, size),
 		w:         math.Exp(math.Log(rand.Float64()) / float64(size)),
 	}
 }
@@ -47,7 +47,7 @@ func (s *ReservoirSampler) Draw() *Diagram {
 
 func (s *ReservoirSampler) Close() {
 	for _, r := range s.reservoir {
-		r.Release()
+		r.r.Release()
 	}
 	s.next.Close()
 }
@@ -71,14 +71,19 @@ func (s *ReservoirSampler) fill(r arrow.Record) arrow.Record {
 	}
 
 	if s.n+r.NumRows() <= s.size { // The record fits in the reservoir
-		s.reservoir = append(s.reservoir, r)
-		r.Retain()
+		for i := int64(0); i < r.NumRows(); i++ {
+			s.reservoir = append(s.reservoir, sample{r: r, i: i})
+			r.Retain()
+		}
 		s.n += r.NumRows()
 		return nil
 	}
 
 	// The record partially fits in the reservoir
-	s.reservoir = append(s.reservoir, r.NewSlice(0, s.size-s.n))
+	for i := int64(0); i < s.size-s.n; i++ {
+		s.reservoir = append(s.reservoir, sample{r: r, i: i})
+		r.Retain()
+	}
 	r = r.NewSlice(s.size-s.n, r.NumRows())
 	s.n = s.size
 	return r
@@ -90,7 +95,7 @@ func (s *ReservoirSampler) sample(r arrow.Record) {
 	if s.i == 0 {
 		s.i = float64(s.n) - 1
 	} else if s.i < float64(n) {
-		s.replace(rand.Intn(int(s.size)), r.NewSlice(int64(s.i)-s.n, int64(s.i)-s.n+1))
+		s.replace(rand.Intn(int(s.size)), sample{r: r, i: int64(s.i) - s.n})
 		s.w = s.w * math.Exp(math.Log(rand.Float64())/float64(s.size))
 	}
 
@@ -98,7 +103,7 @@ func (s *ReservoirSampler) sample(r arrow.Record) {
 		s.i += math.Floor(math.Log(rand.Float64())/math.Log(1-s.w)) + 1
 		if s.i < float64(n) {
 			// replace a random item of the reservoir with row i
-			s.replace(rand.Intn(int(s.size)), r.NewSlice(int64(s.i)-s.n, int64(s.i)-s.n+1))
+			s.replace(rand.Intn(int(s.size)), sample{r: r, i: int64(s.i) - s.n})
 			s.w = s.w * math.Exp(math.Log(rand.Float64())/float64(s.size))
 		}
 	}
@@ -109,7 +114,9 @@ func (s *ReservoirSampler) sample(r arrow.Record) {
 func (s *ReservoirSampler) Finish(ctx context.Context) error {
 	// Send all the records in the reservoir to the next operator
 	for _, r := range s.reservoir {
-		if err := s.next.Callback(ctx, r); err != nil {
+		record := r.r.NewSlice(r.i, r.i+1)
+		defer record.Release()
+		if err := s.next.Callback(ctx, record); err != nil {
 			return err
 		}
 	}
@@ -118,23 +125,13 @@ func (s *ReservoirSampler) Finish(ctx context.Context) error {
 }
 
 // replace will replace the row at index i with the row in the record r at index j.
-func (s *ReservoirSampler) replace(i int, newRow arrow.Record) {
-	// find the record in the reservoir that contains the row at index i
-	rows := int64(0)
-	for k, record := range s.reservoir {
-		if int64(i) < rows+record.NumRows() { // Row i is contained in this record
-			// Edge case where we're replacing a record that is a single row
-			if record.NumRows() == 1 {
-				s.reservoir[k].Release()
-				s.reservoir[k] = newRow
-				return
-			}
-			front := record.NewSlice(0, int64(i)-rows)
-			back := record.NewSlice(int64(i)-rows+1, record.NumRows())
-			s.reservoir[k].Release() // Release the old reference to the entire record
-			s.reservoir = append(s.reservoir[:k], append([]arrow.Record{front, newRow, back}, s.reservoir[k+1:]...)...)
-			return
-		}
-		rows += record.NumRows()
-	}
+func (s *ReservoirSampler) replace(i int, newRow sample) {
+	s.reservoir[i].r.Release()
+	s.reservoir[i] = newRow
+	newRow.r.Retain()
+}
+
+type sample struct {
+	r arrow.Record
+	i int64
 }