@brancz 2024-05-30
So while there are other ways to achieve this, the requirements in Parca/Polar Siganls Cloud that must continue to hold true are:

1) sharing a link must result in users seeing exactly the same result
2) data across visualizations must be consistent (eg. viewing data in an icicle graph must show the same values as tables)
3) the binary names returned from the new metadata call must be consistent with what's shown in the icicle graph

The only way I can see this continuing to be true would be if we cache results from a query indefinitely, and only render reports based on the cached result.